Add "no-refetch-timeout" for StoragePolicy.PRIORITIZE_UPSTREAM_METADATA
This commit adds an option to define a timeout until no re-fetch of
maven-metadata.xml is done.

This gives the possibility to use the local copy until the defined
period is over. Especially on connections with "huge" delays this
setting improves build times and still has a deprecation mechanism
for maven-metadata.xml content.

The default value is 0 seconds (it fetches the file every time).